Minor bugfixes and added option for specifying backup password
* OpenSSH 7.0 and greater similarly disables the ssh-dss (DSA) public key
algorithm. It too is weak and the maintainers recommend against its use.
The README file was updated to use RSA instead of DSA for creating the
backup key.

* Added option to specify a backup password to be used for all backups.
The RouterOS default is to use the password of the user taking the backup.
If using another user but admin the password might be different between
routers and memorizing several passwords isn't' practical enough.